Creating Product Card

+++ src/assets/components/DealsEl.jsx
@@ -7,27 +7,62 @@
 
 const DealsEl=(props)=>{
 
+  let isExisting = false;
+  let { Cart, setCart } = useContext(AuthContext);
+  
+  const { Added, setAdded } = useContext(AuthContext);
+  let { data_from_searver, setData_from_searver } = useContext(AuthContext);
+
+  const { showAdded, setShowAdded } = useContext(AuthContext);
+  if (data_from_searver) {
+    Cart = data_from_searver;
+  }
+   
     
-      let count = 0;
-      let { Cart, setCart } = useContext(AuthContext);
-    
-      const { Added, setAdded } = useContext(AuthContext);
-      let { data_from_searver, setData_from_searver } = useContext(AuthContext);
-    if(data_from_searver){
-      Cart=data_from_searver
+  const CheckingIfItExists = (id, prodpic, title, price) => {
+    const Consists = data_from_searver.some((prev) => prev.id === id);
+
+    if (Consists) {
+      console.log("its here");
+      isExisting = true;
+    } else {
+      console.log("not here");
+      isExisting = false;
     }
-      function CartItem(id, title, price,prodpic) {
-        Cart.push({
-          id: id,
+
+    if (isExisting) {
+      setAdded((prev) => {
+        return {
+          title: "Already Here",
+          price: 0,
+        };
+      });
+    } else {
+      Cart.push({
+        id: id,
+        title: title,
+        price: price,
+        prodpic: prodpic,
+      });
+      localStorage.setItem("Cart", JSON.stringify(Cart));
+
+      setAdded((prev) => {
+        return {
           title: title,
           price: price,
-          prodpic:prodpic
-        });
-        localStorage.setItem("Cart", JSON.stringify(Cart));
-       
-        console.log(data_from_searver);
-       
-      }
+        };
+      });
+    }
+  };
+
+  if (data_from_searver) {
+    Cart = data_from_searver;
+  }
+
+  const CartItem = (id, title, price, prodpic) => {
+    CheckingIfItExists(id, prodpic, title, price);
+    setShowAdded(true);
+  };
     
       return (
         <div className="deal">
